requirements.txt
amqp==2.6.1
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
celery==4.4.2
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.7
django-import-export==2.5.0
et-xmlfile==1.0.1
gunicorn==20.1.0
heroku==0.1.4
idna==2.10
importlib-metadata==3.7.2
kombu==4.6.11
lxml==4.6.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
python-dateutil==1.5
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
soupsieve==2.2
sqlparse==0.4.1
tablib==3.0.0
typing-extensions==3.7.4.3
urllib3==1.26.3
vine==1.3.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.1
